affichage cd

+++ include/Sommet.h
@@ -19,7 +19,7 @@ class Sommet
         int m_x;
         int m_y;
         int m_marquage=0;
-        int m_Cd=0;
+        double m_Cd=0;
         int m_C=0;
         int m_Cvp=1;
     public:
@@ -33,12 +33,12 @@ class Sommet
         int getMarquage() const;
         int getC()const;
         int getCvp()const;
-        int getCd()const;
+        double getCd()const;
         std::string getNom () const;
         void Adjacent(int ID);
         void Erase_Adjacent();
         void setMarquage(int coul);
-        void setCd(int C);
+        void setCd(double C);
         void setCvp(int C);
         void setC(int C);
         void Adjtrie();
